+++ source/src/RTSCameraAgentComponent.cs
@@ -1,4 +1,5 @@
-﻿using RTSCamera.QuerySystem;
+﻿using System;
+using RTSCamera.QuerySystem;
 using TaleWorlds.Engine;
 using TaleWorlds.Library;
 using TaleWorlds.MountAndBlade;
@@ -55,55 +56,75 @@ namespace RTSCamera
                 var targetFormation = QueryDataStore.Get(formation.TargetFormation);
 
                 Vec2 unitPosition;
+                if (QueryLibrary.IsRangedCavalry(unit))
+                {
+                    unitPosition = unit.Position.AsVec2;
+                    return targetFormation
+                        .NearestAgent(unitPosition)?.GetWorldPosition() ?? new WorldPosition();
+                }
                 if (QueryLibrary.IsCavalry(unit))
                 {
-                    if (QueryLibrary.IsRangedCavalry(unit))
+                    unitPosition = formation.GetCurrentGlobalPositionOfUnit(unit, true) * 0.2f +
+                                   unit.Position.AsVec2 * 0.8f;
+                    var targetAgent = targetFormation.NearestOfAverageOfNearestPosition(unitPosition, 7);
+                    if (targetAgent != null)
                     {
-                        unitPosition = unit.Position.AsVec2;
-                        return targetFormation
-                            .NearestAgent(unitPosition)?.GetWorldPosition() ?? new WorldPosition();
-                    }
-                    else
-                    {
-                        unitPosition = formation.GetCurrentGlobalPositionOfUnit(unit, true) * 0.2f +
-                                       unit.Position.AsVec2 * 0.8f;
-                        var targetPosition = targetFormation
-                            .NearestOfAverageOfNearestPosition(unitPosition, 10)?.GetWorldPosition();
-                        if (targetPosition != null)
-                        {
-                            var result  = targetPosition.Value;
-                            var targetDirection = result.AsVec2 - unit.Position.AsVec2;
-                            var distance = targetDirection.Normalize();
-                            if (distance < 3)
-                            {
-                                result = unit.GetWorldPosition();
-                                result.SetVec2(CurrentDirection * 20 + result.AsVec2);
-                            }
-                            else
-                            {
-                                if (distance < 20 && targetDirection.DotProduct(CurrentDirection) < 0)
-                                {
-                                    result.SetVec2(-targetDirection * 50 + result.AsVec2);
-                                }
-                                else
-                                {
-                                    CurrentDirection = targetDirection;
-                                    result.SetVec2(targetDirection * 10 + result.AsVec2);
-                                }
-                            }
+                        if (targetAgent.HasMount)
+                            return targetAgent.GetWorldPosition();
 
-                            return result;
+                        var targetPosition = targetAgent.GetWorldPosition();
+                        var targetDirection = targetPosition.AsVec2 - unit.Position.AsVec2;
+                        var distance = targetDirection.Normalize();
+                        var result = targetPosition;
+
+                        // new
+                        if (distance > 20)
+                        {
+                            CurrentDirection = targetDirection;
+                            result.SetVec2(targetDirection * 5 + targetPosition.AsVec2);
+                        }
+                        else if (targetDirection.DotProduct(CurrentDirection) < 0)
+                        {
+                            result.SetVec2((CurrentDirection.DotProduct(targetDirection * distance) + 50) * CurrentDirection + unit.Position.AsVec2);
+                        }
+                        else
+                        {
+                            result.SetVec2(CurrentDirection * 5 + targetPosition.AsVec2);
                         }
 
-                        return new WorldPosition();
+
+                        // old
+                        //if (distance < 3)
+                        //{
+                        //    result = unit.GetWorldPosition();
+                        //    result.SetVec2(CurrentDirection * 20 + result.AsVec2);
+                        //}
+                        //else
+                        //{
+                        //    if (distance < 20 && targetDirection.DotProduct(CurrentDirection) < 0)
+                        //    {
+                        //        result.SetVec2(-targetDirection * 50 + result.AsVec2);
+                        //    }
+                        //    else
+                        //    {
+                        //        CurrentDirection = targetDirection;
+                        //        result.SetVec2(targetDirection * 10 + result.AsVec2);
+                        //    }
+
+                        //}
+
+
+                        return result.GetNavMesh() == UIntPtr.Zero || !Mission.Current.IsPositionInsideBoundaries(result.AsVec2) ? targetPosition : result;
                     }
+
+                    return new WorldPosition();
                 }
 
                 unitPosition = formation.GetCurrentGlobalPositionOfUnit(unit, true) * 0.2f +
                                unit.Position.AsVec2 * 0.8f;
                 return targetFormation
                     .NearestAgent(unitPosition)?.GetWorldPosition() ?? new WorldPosition();
-            }, 0.1f);
+            }, 0.2f);
         }
 
         public void SetContourColor(int level, uint? color, bool alwaysVisible)
@@ -113,6 +134,11 @@ namespace RTSCamera
                 _currentLevel = color.HasValue ? level : EffectiveLevel(level - 1);
                 SetColor();
             }
+        }
+
+        protected override void OnTickAsAI(float dt)
+        {
+            base.OnTickAsAI(dt);
         }
 
         public bool SetContourColorWithoutUpdate(int level, uint? color, bool alwaysVisible)
@@ -145,8 +171,8 @@ namespace RTSCamera
 
         public void ClearTargetOrSelectedFormationColor()
         {
-            bool needUpdate = SetContourColorWithoutUpdate((int) ColorLevel.TargetFormation, null, true);
-            needUpdate |= SetContourColorWithoutUpdate((int) ColorLevel.SelectedFormation, null, true);
+            bool needUpdate = SetContourColorWithoutUpdate((int)ColorLevel.TargetFormation, null, true);
+            needUpdate |= SetContourColorWithoutUpdate((int)ColorLevel.SelectedFormation, null, true);
             if (needUpdate)
                 UpdateColor();
         }